requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
langchain
langchain-google-genai
